Crate custom_data_loader

This example demonstrates how to implement and register a re_data_loader::DataLoader into the Rerun Viewer in order to add support for loading arbitrary files.

Usage:

$ cargo r -p custom_data_loader -- path/to/some/file

Structs

HashLoader 🔒

A custom re_data_loader::DataLoader that logs the hash of file as a rerun::TextDocument.

Functions

hash_and_log 🔒

main 🔒